refactor(StationRequest,StationResponse): move to dto package
