+++ src/main/java/subway/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.dto;
 
 public class StationResponse {
     private Long id;
